Don't call shutdown in a trap context as it breaks Celluloid.
Previously, shutdown was called directly from trap(:INT) but this meant that practically speaking the function was not very useful. Using the self-pipe trick was common to notify the main daemon process that it needed to complete it's work. This is now build directly into process-daemon.

There are two modes of operation.

1/ The daemon startup function never exits and internally calls something like IO.select - when SIGINT is received, ruby would raise the Interrupt exception and it would propagage up the stack normally and terminate the daemon.

2/ The daemon startup function starts up some threads/actors and then sleeps. Rather than calling sleep, call sleep_until_interrupted which will simply wait until SIGINT is received and then proceed to call the shutdown function.

As usual, a daemon can simply exit from the startup function. this is considered a normal termination of the daemon.